2 commits in einem (Keine Zeit die beiden zu trennen sowy) Wenn man auf ein "Pay Feld" kommt wird ein das gleiche Fenster wie beim Straßen kaufen angezeigt. Da wird man dann aufgefordert das Geld zu zahlen. Wenn man dann auf Zahlen drückt und es nicht funktioniert gibt es eine Error animation. Überhaupt wurde dem ganzem Fenster animation hinzugefügt. Frei Parken funktioniert Gefängnis Funktioniert

+++ src/main/java/sandwich/de/monopoly/Fields/Station.java
@@ -19,7 +19,7 @@ public class Station extends Field {
     private final String name;
     private final double price;
 
-    public Station(String name ,double price, double postion) {
+    public Station(String name ,int price, double postion) {
         super(postion);
         this.name = name;
         this.price = price;
